requirements.txt
joblib==1.4.2
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.3.2
scikit-surprise==1.1.4
scipy==1.10.1
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
